refactor: form utils cleanup
