--- packages/dm-core-plugins/src/form/components/StyledComponents.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-export const VisibleOnHover = styled.div`
-  visibility: hidden;
-  &:hover {
-    visibility: visible;
-  }
-`
