+++ packages/dm-core-plugins/src/form/utils/complexEntityUtils.ts
@@ -1,4 +1,4 @@
-import { TFormConfig, TUiAttributeObject } from '../../types'
+import { TFormConfig, TUiAttributeObject } from '../types'
 
 export const getExpandViewConfig = (uiAttribute?: TUiAttributeObject) => {
   return (
